+++ frontend/src/App.jsx
@@ -1,5 +1,5 @@
 import Content from "./components/Content";
-import Nav from "./components/Nav";
+import Nav from "./components/Navbar";
 
 function Home() {
   return (
